reset value at vertex method change
If the user would select a different vertex selection method, the
constraint values would revert to "no". This is naively safe in the
sense that it wouldn't over constraint the model.
Practically it would mean the model would fail and you spend minutes to
hours figuring out what it is and than it's this. Now it will give a
proper error.